+++ database/migrations/2022_11_15_220906_create_groups_table.php
@@ -15,7 +15,7 @@ class CreateGroupsTable extends Migration
     {
         Schema::create('groups', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('owner');
+            $table->unsignedBigInteger('owner');
             $table->string('name');
             $table->string('description');
             $table->string('photo');
